Make makeFetchWithTimeout more usable
* move makeFetchWithTimeout to utils
* add a new parameter for a timeout in fetchWithTimeout
* update uns fetching functions
