+++ background/services/name/resolvers/uns.ts
@@ -1,31 +1,12 @@
 import { AddressOnNetwork, NameOnNetwork } from "../../../accounts"
 import { ETHEREUM, POLYGON, SECOND } from "../../../constants"
-import logger from "../../../lib/logger"
 import { isDefined } from "../../../lib/utils/type-guards"
 import { sameNetwork } from "../../../networks"
+import fetchWithTimeout from "../../../utils/fetching"
 import { NameResolver } from "../name-resolver"
 
 // Time until response is stale
 const RESPONSE_TTL = 5 * SECOND
-
-const makeFetchWithTimeout = (timeoutMs: number) => {
-  return async function fetchWithTimeout(
-    requestInfo: RequestInfo,
-    options?: RequestInit | undefined
-  ) {
-    const controller = new AbortController()
-    const id = setTimeout(() => {
-      logger.warn("Request to ", requestInfo, " timed out")
-      return controller.abort()
-    }, timeoutMs)
-    const response = await fetch(requestInfo, {
-      ...options,
-      signal: controller.signal,
-    })
-    clearTimeout(id)
-    return response
-  }
-}
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 const cacheAsyncResults = <F extends (...args: any[]) => Promise<any>>(
@@ -80,7 +61,6 @@ const cacheAsyncResults = <F extends (...args: any[]) => Promise<any>>(
     return result
   }) as F
 }
-const fetchWithTimeout = makeFetchWithTimeout(3_000)
 
 const UNS_SUPPORTED_NETWORKS = [ETHEREUM, POLYGON]
 
@@ -88,7 +68,7 @@ const UNS_SUPPORTED_NETWORKS = [ETHEREUM, POLYGON]
  * Lookup a UNS domain name and fetch the owners address
  */
 const lookupUNSDomain = async (domain: string) => {
-  const response = await fetchWithTimeout(
+  const response = await fetchWithTimeout()(
     `https://unstoppabledomains.g.alchemy.com/domains/${domain}`,
     {
       method: "GET",
@@ -108,7 +88,7 @@ const lookupUNSDomain = async (domain: string) => {
 const reverseLookupAddress = cacheAsyncResults(
   RESPONSE_TTL,
   async (address: string) => {
-    const response = await fetchWithTimeout(
+    const response = await fetchWithTimeout()(
       `https://unstoppabledomains.g.alchemy.com/domains/?owners=${address}&sortBy=id&sortDirection=ASC`,
       {
         method: "GET",
